+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.65](https://github.com/realityforge/proton/tree/v0.65) (2025-01-13) · [Full Changelog](https://github.com/realityforge/proton/compare/v0.64...v0.65)
 
 Changes in this release: